% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASToolClusterPckg-package.R
\docType{package}
\name{CASToolClusterPckg-package}
\alias{CASToolClusterPckg}
\alias{CASToolClusterPckg-package}
\title{CASToolClusterPckg: Serves cluster assignments, cluster figures, and state boundaries for the CASTool}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Laura Naslund \email{Naslund.Laura@epa.gov}

}
\keyword{internal}
